Instagram Fake Profile Detector/requirements.txt
pandas
scikit-learn
joblib
instaloader
streamlit
torch
torchvision
Pillow
